+++ packages/vue-cli-plugin-vue-dash/generator/template/src/theme/colors.ts
@@ -1,7 +1,7 @@
 // Import Vuetify theme from Design Tokens
-import { lightTheme } from '@cnamts/design-tokens/src/colors';
+import { colorTheme } from '@cnamts/design-tokens/src/colors';
 
 /** Custom Vuetify color theme */
 export const themeColors = {
-	light: lightTheme
+	light: colorTheme
 };